+++ addon/plugins/variable-plugin/index.ts
@@ -1,5 +1,0 @@
-export type TemplateVariablePluginOptions = {
-  endpoint: string;
-  zonalLocationCodelistUri: string;
-  nonZonalLocationCodelistUri: string;
-};